Adds the ability to rewrite responses from the AnnotationManagerBean
