+++ core/query/pom.xml
@@ -66,6 +66,12 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <!-- JAX-RS API for ResponseRewriter interface (javax.ws.rs.core.Response) -->
+        <dependency>
+            <groupId>org.jboss.spec.javax.ws.rs</groupId>
+            <artifactId>jboss-jaxrs-api_2.1_spec</artifactId>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
     <build>
         <plugins>
