+++ web-services/annotations/src/main/java/datawave/webservice/annotation/AnnotationManagerBean.java
@@ -2,7 +2,6 @@ package datawave.webservice.annotation;
 
 import static datawave.annotation.util.v1.AnnotationUtils.injectAnnotationSource;
 
-import java.security.Principal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -10,9 +9,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.ExecutorService;
 
 import javax.annotation.Resource;
 import javax.annotation.security.DeclareRoles;
@@ -33,26 +29,19 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import org.apache.accumulo.core.client.AccumuloClient;
-import org.apache.accumulo.core.security.Authorizations;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.annotations.VisibleForTesting;
 
-import datawave.annotation.data.v1.AccumuloAnnotationSerializer;
-import datawave.annotation.data.v1.AccumuloAnnotationSourceSerializer;
-import datawave.annotation.data.v1.AnnotationDataAccess;
 import datawave.annotation.data.v1.AnnotationReader;
-import datawave.annotation.data.v1.FederatedAnnotationReader;
 import datawave.annotation.protobuf.v1.Annotation;
 import datawave.annotation.protobuf.v1.AnnotationSource;
 import datawave.annotation.protobuf.v1.Segment;
 import datawave.configuration.spring.SpringBean;
 import datawave.core.common.connection.AccumuloConnectionFactory;
-import datawave.microservice.authorization.util.AuthorizationsUtil;
-import datawave.security.authorization.DatawavePrincipal;
+import datawave.core.query.logic.ResponseRewriter;
 import datawave.webservice.query.exception.QueryException;
 import datawave.webservice.query.result.event.ResponseObjectFactory;
 import datawave.webservice.query.runner.AccumuloConnectionRequestBean;
@@ -85,6 +74,10 @@ public class AnnotationManagerBean implements AnnotationManager {
     @SpringBean(name = "AnnotationManagerConfig")
     private AnnotationManagerConfig config;
 
+    @Inject
+    @SpringBean(name = "AnnotationResponseRewriter", required = false)
+    private ResponseRewriter responseRewriter;
+
     @Resource
     private ManagedExecutorService annotationFederatedReadExecutor;
 
@@ -98,14 +91,14 @@ public class AnnotationManagerBean implements AnnotationManager {
     @Produces("application/json")
     @Override
     public Response getAnnotationSource(@PathParam("analyticHash") String analyticHash) {
-        final RequestContext context = newRequestContext();
+        final AnnotationManagerRequestContext context = newRequestContext();
         try {
             context.initializeAnnotationService();
             Optional<AnnotationSource> results = context.getAnnotationSource(analyticHash);
             if (results.isEmpty()) {
                 return jsonNotFound("No annotation source found for analyticHash: " + analyticHash);
             }
-            return jsonOk(results.get());
+            return jsonOk(results.get(), context);
         } catch (Exception e) {
             final String message = String.format("Internal error fetching annotation source: %s", e.getMessage());
             log.error(message, e);
@@ -121,7 +114,7 @@ public class AnnotationManagerBean implements AnnotationManager {
     @Override
     public Response getAnnotationTypes(@PathParam("idType") String idType, @PathParam("id") String id) {
         // TODO sanitize input to make sure it contains nothing weird like nulls.
-        final RequestContext context = newRequestContext();
+        final AnnotationManagerRequestContext context = newRequestContext();
         try {
             final List<Metadata> metadata = lookupDocumentIdentifier(context, idType, id);
             if (metadata.isEmpty()) {
@@ -138,7 +131,7 @@ public class AnnotationManagerBean implements AnnotationManager {
             if (results.isEmpty()) {
                 return jsonNotFound("annotation types", idType, id, metadata.toString(), null, null, null);
             }
-            return jsonOk(results);
+            return jsonOk(results, context);
         } catch (Exception e) {
             final String message = String.format("Internal error fetching annotation: %s", e.getMessage());
             log.error(message, e);
@@ -154,7 +147,7 @@ public class AnnotationManagerBean implements AnnotationManager {
     @Override
     public Response getAnnotationsFor(@PathParam("idType") String idType, @PathParam("id") String id) {
         // TODO sanitize input to make sure it contains nothing weird like nulls.
-        final RequestContext context = newRequestContext();
+        final AnnotationManagerRequestContext context = newRequestContext();
         try {
             final List<Metadata> metadata = lookupDocumentIdentifier(context, idType, id);
             if (metadata.isEmpty()) {
@@ -173,7 +166,7 @@ public class AnnotationManagerBean implements AnnotationManager {
             if (results.isEmpty()) {
                 return jsonNotFound("annotations", idType, id, metadata.toString(), null, null, null);
             }
-            return jsonOk(results);
+            return jsonOk(results, context);
         } catch (Exception e) {
             final String message = String.format("Internal error fetching annotation: %s", e.getMessage());
             log.error(message, e);
@@ -189,7 +182,7 @@ public class AnnotationManagerBean implements AnnotationManager {
     @Override
     public Response getAnnotationsByType(@PathParam("idType") String idType, @PathParam("id") String id, @PathParam("annotationType") String annotationType) {
         // TODO sanitize input to make sure it contains nothing weird like nulls.
-        final RequestContext context = newRequestContext();
+        final AnnotationManagerRequestContext context = newRequestContext();
         try {
             final List<Metadata> metadata = lookupDocumentIdentifier(context, idType, id);
             if (metadata.isEmpty()) {
@@ -209,7 +202,7 @@ public class AnnotationManagerBean implements AnnotationManager {
             if (results.isEmpty()) {
                 return jsonNotFound("annotations of type", idType, id, metadata.toString(), annotationType, null, null);
             }
-            return jsonOk(results);
+            return jsonOk(results, context);
         } catch (Exception e) {
             final String message = String.format("Internal error fetching annotation: %s", e.getMessage());
             log.error(message, e);
@@ -224,7 +217,7 @@ public class AnnotationManagerBean implements AnnotationManager {
     @Produces("application/json")
     @Override
     public Response getAnnotation(@PathParam("idType") String idType, @PathParam("id") String id, @PathParam("annotationId") String annotationId) {
-        final RequestContext context = newRequestContext();
+        final AnnotationManagerRequestContext context = newRequestContext();
         try {
             final List<Metadata> metadata = lookupDocumentIdentifier(context, idType, id);
             if (metadata.isEmpty()) {
@@ -243,7 +236,7 @@ public class AnnotationManagerBean implements AnnotationManager {
             if (results.isEmpty()) {
                 return jsonNotFound("annotations", idType, id, metadata.toString(), null, annotationId, null);
             }
-            return jsonOk(results);
+            return jsonOk(results, context);
         } catch (Exception e) {
             final String message = String.format("Internal error fetching annotation: %s", e.getMessage());
             log.error(message, e);
@@ -259,7 +252,7 @@ public class AnnotationManagerBean implements AnnotationManager {
     @Override
     public Response getAnnotationSegment(@PathParam("idType") String idType, @PathParam("id") String id, @PathParam("annotationId") String annotationId,
                     @PathParam("segmentHash") String segmentHash) {
-        final RequestContext context = newRequestContext();
+        final AnnotationManagerRequestContext context = newRequestContext();
         try {
             final List<Metadata> metadata = lookupDocumentIdentifier(context, idType, id);
             if (metadata.isEmpty()) {
@@ -294,7 +287,7 @@ public class AnnotationManagerBean implements AnnotationManager {
             if (results.isEmpty()) {
                 return jsonNotFound("segments", idType, id, metadata.toString(), null, annotationId, segmentHash);
             }
-            return jsonOk(results);
+            return jsonOk(results, context);
         } catch (QueryException e) {
             final String message = String.format("Internal error fetching segment: %s", e.getMessage());
             log.error(message, e);
@@ -316,7 +309,7 @@ public class AnnotationManagerBean implements AnnotationManager {
      * @throws QueryException
      *             if the id is malformed.
      */
-    private List<Metadata> lookupDocumentIdentifier(RequestContext context, String idType, String id) throws QueryException {
+    private List<Metadata> lookupDocumentIdentifier(AnnotationManagerRequestContext context, String idType, String id) throws QueryException {
         // If the idType is RECORD_ID or DOCUMENT, treat the id provided as an internal id and perform a direct lookup
         // against the annotations table, if that's enabled.
         if (idType.equals("DOCUMENT") || idType.equals("RECORD_ID")) {
@@ -341,7 +334,7 @@ public class AnnotationManagerBean implements AnnotationManager {
      *            the annotations to inject sources into
      * @return return annotations with sources injected where possible.
      */
-    private List<Annotation> lookupAndInjectAnnotationSources(RequestContext context, Collection<Annotation> annotations) {
+    private List<Annotation> lookupAndInjectAnnotationSources(AnnotationManagerRequestContext context, Collection<Annotation> annotations) {
         final List<Annotation> results = new ArrayList<>();
         for (Annotation a : annotations) {
             results.add(lookupAndInjectAnnotationSource(context, a));
@@ -353,7 +346,7 @@ public class AnnotationManagerBean implements AnnotationManager {
      * Given an annotation, retrieve the annotation source information that is referenced by their analyticHash. Employs a per-request hash so we don't look up
      * a single source multiple times.
      */
-    private Annotation lookupAndInjectAnnotationSource(RequestContext context, Annotation a) {
+    private Annotation lookupAndInjectAnnotationSource(AnnotationManagerRequestContext context, Annotation a) {
         // no need to inject a source if we already have one.
         if (a.hasSource()) {
             log.warn("Strange, this annotation already has a source. Annotation {}/{}/{} {}, using analyticHash {}", a.getShard(), a.getDataType(), a.getUid(),
@@ -461,239 +454,38 @@ public class AnnotationManagerBean implements AnnotationManager {
         return Response.ok(responseObject, MediaType.APPLICATION_JSON_TYPE.withCharset("utf-8")).build();
     }
 
+    /**
+     * Apply the response rewriter (if configured) to transform the response before returning it to the client. Only applies to successful (200 OK) responses.
+     * Error responses are returned as-is.
+     *
+     * @param responseObject
+     *            the response object
+     * @param context
+     *            the request context for the rewriter
+     * @return the rewritten response, or the original if no rewriter is configured
+     */
+    Response jsonOk(Object responseObject, AnnotationManagerRequestContext context) {
+        Response response = jsonOk(responseObject);
+        if (responseRewriter != null) {
+            try {
+                return responseRewriter.rewriteResponse(response, context);
+            } catch (Exception e) {
+                final String message = String.format("Error applying response rewriter: %s", e.getMessage());
+                log.error(message, e);
+                return jsonError(message);
+            }
+        }
+        return response;
+    }
+
     @VisibleForTesting
     protected AnnotationManagerConfig getConfig() {
         return config;
     }
 
-    private RequestContext newRequestContext() {
-        return new RequestContext(config, ctx, connectionFactory, accumuloConnectionRequestBean, responseObjectFactory, annotationFederatedReadExecutor);
+    private AnnotationManagerRequestContext newRequestContext() {
+        return new AnnotationManagerRequestContext(config, ctx, connectionFactory, accumuloConnectionRequestBean, responseObjectFactory,
+                        annotationFederatedReadExecutor);
     }
 
-    /** Per-request initialization code and related state */
-    protected static final class RequestContext {
-
-        private final AnnotationManagerConfig config;
-        private final AccumuloConnectionFactory connectionFactory;
-        private final AccumuloConnectionRequestBean accumuloConnectionRequestBean;
-        private final ResponseObjectFactory responseObjectFactory;
-        private final ExecutorService annotationFederatedReadExecutor;
-
-        /** the user performing this request */
-        private final DatawavePrincipal datawavePrincipal;
-
-        /** the dn of the user performing this request */
-        private final String userDn;
-
-        /** proxy servers involved in this request */
-        private final Collection<String> proxyServers;
-
-        /** quthorizations pulled from the query parameters for this request */
-        private final String queryAuths;
-
-        /** the final set of merged query and user authorizations. */
-        private Set<Authorizations> authorizations;
-
-        /** the accumulo client to use for this request - obtained from the connection pool and must be returned. */
-        private AccumuloClient client;
-
-        /** used to lookup uuids and obtain internal identifiers, scoped to the caller's authorizations */
-        private LookupUUIDService lookupUUIDService;
-
-        /** used to _read_ annotations directly from accumulo, scoped to the caller's authorizations */
-        private AnnotationReader annotationDataAccess;
-
-        /** Cache lookups for unique analytic source hashes so we don't perform lookups more than once. TODO: make this a proper cross-request cache? */
-        private final Map<String,Optional<AnnotationSource>> retrievedSourcesCache = new HashMap<>();
-
-        /** Lookup an annotation source or retrieve it from the cache */
-        public Optional<AnnotationSource> getAnnotationSource(String analyticHash) {
-            return retrievedSourcesCache.computeIfAbsent(analyticHash, key -> annotationDataAccess.getAnnotationSource(analyticHash));
-        }
-
-        /**
-         * Initialize the request context with the objects needed to perform various request state initialization. Validation of the objects provided is
-         * performed in the various initialize methods exposed by this class. Each of the objects provided as parameters are expected to be shared across many
-         * requests.
-         *
-         * @param config
-         *            the annotation manager configuration
-         * @param ctx
-         *            the ejb context - used for retrieving the principal for he qrequest
-         * @param connectionFactory
-         *            the accumulo connection factory - used for getting accumulo clients
-         * @param accumuloConnectionRequestBean
-         *            the accumulo connection request bean - used for tracking accumulo clients rerquests
-         * @param responseObjectFactory
-         *            the response object factory used for creating LookupUUID responses.
-         */
-        protected RequestContext(AnnotationManagerConfig config, EJBContext ctx, AccumuloConnectionFactory connectionFactory,
-                        AccumuloConnectionRequestBean accumuloConnectionRequestBean, ResponseObjectFactory responseObjectFactory,
-                        ExecutorService annotationFederatedReadExecutor) {
-            this.config = config;
-            this.connectionFactory = connectionFactory;
-            this.accumuloConnectionRequestBean = accumuloConnectionRequestBean;
-            this.responseObjectFactory = responseObjectFactory;
-            this.annotationFederatedReadExecutor = annotationFederatedReadExecutor;
-
-            final Principal p = ctx.getCallerPrincipal();
-            final boolean isDatawavePrincipal = DatawavePrincipal.class.isAssignableFrom(p.getClass());
-            final DatawavePrincipal dp = isDatawavePrincipal ? (DatawavePrincipal) p : null;
-
-            this.userDn = dp != null ? dp.getUserDN().subjectDN() : p.getName();
-            this.proxyServers = dp != null ? dp.getProxyServers() : null;
-            this.datawavePrincipal = dp;
-
-            // TODO: allow downgrading by reading query auths from query parameters.
-            this.queryAuths = null;
-        }
-
-        /**
-         * Calculate the authorizations for this request based on the principal and the queryAuths if any.
-         *
-         * @return a valid set of query auths, will throw an exception if this isn't possible.
-         * @throws QueryException
-         *             and exception if there was a problem calculating the auths
-         */
-        private Set<Authorizations> initializeAuthorizations() throws QueryException {
-            if (authorizations == null) {
-                log.trace("Initializing authorizations: userDn: {}, query: {}", userDn, queryAuths);
-                if (datawavePrincipal == null) {
-                    throw new QueryException("Failed to get user principal from request, unable to proceed");
-                }
-
-                try {
-                    if (queryAuths == null) {
-                        authorizations = AuthorizationsUtil.buildAuthorizations(datawavePrincipal.getAuthorizations());
-                    } else {
-                        final String downgradedAuths = AuthorizationsUtil.downgradeUserAuths(queryAuths, datawavePrincipal, datawavePrincipal);
-                        authorizations = AuthorizationsUtil.buildAuthorizations(Collections.singleton(AuthorizationsUtil.splitAuths(downgradedAuths)));
-                    }
-
-                } catch (Exception e) {
-                    throw new QueryException("Failed to get user query authorizations", e);
-                }
-
-                log.debug("Authorizations initialized: userDn: {}, query: {}, final auths: {}", userDn, queryAuths, authorizations);
-            }
-            return authorizations;
-        }
-
-        /**
-         * Initialize the accumulo client
-         *
-         * @return a valid client, will throw an exception if this isn't possibly
-         * @throws QueryException
-         *             if the client can't be initialized.
-         */
-        protected AccumuloClient initializeAccumuloClient() throws QueryException {
-            if (client == null) {
-                log.trace("Initializing accumulo client");
-                UUID transactionUUID = java.util.UUID.randomUUID();
-
-                if (connectionFactory == null) {
-                    throw new QueryException("The accumulo connection factory isn't present, unable to proceed");
-                }
-
-                Map<String,String> trackingMap = connectionFactory.getTrackingMap(Thread.currentThread().getStackTrace());
-                if (trackingMap != null) {
-                    trackingMap.put("query.user", userDn);
-                    trackingMap.put("query.id", transactionUUID.toString());
-                    trackingMap.put("query.query", "annotation manager");
-                } else {
-                    log.info("Accumulo connection tracking map was null, this isn't fatal, but odd.");
-                }
-
-                if (accumuloConnectionRequestBean == null) {
-                    throw new QueryException("The accumulo connection request manager isn't present, unable to proceed");
-                }
-
-                accumuloConnectionRequestBean.requestBegin(transactionUUID.toString(), userDn, trackingMap);
-                try {
-                    client = connectionFactory.getClient(userDn, proxyServers, config.getConnPoolName(), config.getPriority(), trackingMap);
-                } catch (Exception e) {
-                    throw new QueryException("Unable to get Accumulo client, exception encountered: ", e);
-                } finally {
-                    accumuloConnectionRequestBean.requestEnd(transactionUUID.toString());
-                }
-                log.debug("Accumulo client initialized successfully");
-
-            }
-            return client;
-        }
-
-        /**
-         * Initialize the lookup uuid service
-         *
-         * @return a valid lookup uuid service, throws an exception if this isn't possible.
-         * @throws QueryException
-         *             if the lookup uuid service can't be initialized.
-         */
-        protected LookupUUIDService initializeLookupUUIDService() throws QueryException {
-            if (lookupUUIDService == null) {
-
-                if (config == null) {
-                    throw new QueryException("The lookup uuid service configuration isn't present, unable to proceed");
-                }
-
-                if (responseObjectFactory == null) {
-                    throw new QueryException("The response object factory isn't present, unable to proceed");
-                }
-
-                log.trace("Initializing lookupUUIDService");
-                final Set<Authorizations> authorizations = initializeAuthorizations();
-                final AccumuloClient client = initializeAccumuloClient();
-                lookupUUIDService = new LookupUUIDService(config.getLookupUUIDServiceConfig(), client, authorizations, responseObjectFactory,
-                                config.getLookupUUIDQueryLogic());
-                log.debug("LookupUUID service initialized successfully");
-            }
-            return lookupUUIDService;
-        }
-
-        /**
-         * Initialize the annotation service, specifically the data access layer.
-         *
-         * @return a valid annotation data access object, throws an exception if this isn't possible.
-         * @throws QueryException
-         *             if the annotation data access object can't be initialized.
-         */
-        protected AnnotationReader initializeAnnotationService() throws QueryException {
-            if (annotationDataAccess == null) {
-                log.trace("Initializing annotation data access layer");
-                final Set<Authorizations> authorizations = initializeAuthorizations();
-                final AccumuloClient client = initializeAccumuloClient();
-                final AccumuloAnnotationSerializer annotationSerializer = new AccumuloAnnotationSerializer(
-                                config.getAnnotationConfig().getVisibilityTransformer(), config.getAnnotationConfig().getTimestampTransformer());
-                final AccumuloAnnotationSourceSerializer annotationSourceSerializer = new AccumuloAnnotationSourceSerializer(
-                                config.getAnnotationConfig().getVisibilityTransformer(), config.getAnnotationConfig().getTimestampTransformer());
-
-                final AnnotationReader annotationReader = new AnnotationDataAccess(client, authorizations,
-                                config.getAnnotationConfig().getAnnotationTableName(), config.getAnnotationConfig().getAnnotationSourceTableName(),
-                                annotationSerializer, annotationSourceSerializer);
-                final AnnotationReader truthmarkReader = new AnnotationDataAccess(client, authorizations, config.getAnnotationConfig().getTruthmarkTableName(),
-                                config.getAnnotationConfig().getTruthmarkSourceTableName(), annotationSerializer, annotationSourceSerializer);
-                Map<String,AnnotationReader> annotationReaderMap = Map.of("annotation", annotationReader, "truthmark", truthmarkReader);
-                annotationDataAccess = new FederatedAnnotationReader(annotationReaderMap, annotationFederatedReadExecutor);
-
-                log.debug("Annotation data access layer initialized successfully");
-            }
-            return annotationDataAccess;
-        }
-
-        /**
-         * Return the accumulo client currently held by this class. If there's a problem returning the client, logs a warning. Sets the internal client state to
-         * null.
-         */
-        protected void returnAccumuloClient() {
-            try {
-                log.trace("Returning accumulo client");
-                connectionFactory.returnClient(client);
-                log.debug("Accumulo client returned");
-            } catch (Exception e) {
-                log.warn("Error when returning client", e);
-            } finally {
-                client = null;
-            }
-        }
-    }
 }
